+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Fixed
+- Fixed bundled font selections such as GenEi Kiwami Gothic not being reflected in rendered videos.
+- Fixed portrait videos clipping wide horizontal subtitles by wrapping lines to fit within 90% of the video width.
+
 ## [1.0.0] - 2026-05-16
 
 ### Added